Move "common"/"internal" targets to top-level cmake
Also removes the global `include_directories` in favour of
target-specific includes which is generally the preferred cmake
approach.

Before the commit the common/internal targets were constructed in src's
CMakeLists, but that is weird because src has to reach outside its own
directory to find include paths, and then `tests` are reaching into a
sibling for the targets they tests need.

The ../ also causes issues with some log statements because, depending
on how a header gets included, might end up being something like
"src/../include/whatever" in the filename which doesn't get stripped
because of the relative path backtracking.