Make dgram-speedtest-server accept multiple connections
Currently dgram-speed-server is annoying because it terminates after
each connection.  This makes it accept multiple connections (like
speedtest-server) which makes testing against it multiple times much
less annoying.

This also changes the final packet send back at the end of the test by
the server to include the number of received packets so that the client
can display that information at the end of the test (previously only the
server displayed it).