Add public `close()` method for app-specified shutdown
This renames the close_gracefully() method to just `close()` and makes
it public so that an application has a bit more control over Network
shutdown.

libsession-util ran into an issue where the quic::Network destruction
segfaults because it has callbacks (e.g. connection closing) that expect
to be able to use their local `quic::Network` instance, but can't
because it's in the process of being destroyed.

Essentially the pattern is this:

    class A {
        quic::Network net;
    };
    A a;
    // start a connection with a connection_close callback that
    // references `a`.

If the quic::Network destruction is triggered by the destruction of `a`,
then the quic::Network destructor is firing *after* `a` is no longer in
a valid state, and so the callback referencing `a` is invalid.

Exposing close() allows the application to deal with this: it can add a
`net.close()` either in `A`'s destructor or some other shutdown code so
that it can ensure the quic Network shutdown happens *before* `a`
becomes unusable.

This also adds a non-synchronous `call_soon()` to mirror the call(ep)
and call_soon(ep) methods for shutdown down individual endpoints.