Fix drop_connection bugs
The way we were dropping connections was not right: these always need to
be deferred into the event loop, because they are called from within a
Connection itself where dropping is not safe.  We had some called that
wrapped this in a `call()`, but those weren't quite right either because
call() fires immediately when already inside the event loop, as these
would be.  This thus makes `drop_connection` itself do the call_soon so
that we can safely use it directly, and renames the existing direct,
unsafe one to `_drop_connection`.

This also adds a missing set_closing(): without it, a drop_connection
can trigger a second attempt to close or drain the connection, which
results in an erroneous double-fire of the close callback and a
potential segfault from entering code multiple times that we shouldn't
be.