Fix wrong-destruction-order crash in callback_waiter
If a callback waiter converted function outlives the callback_waiter
then the callback and promise fulfilment could happen after the
callback_waiter has been destroyed, leading to weird crashes, particular
when the function exits via failure.

(For example, my case that triggered this: a failed REQUIRE() that
started to stack unwind the function and because callback_waiter was
defined *after* the connection and network in my code, the callback that
eventually fired during Network destruction was firing on a callback
with a bad callback_waiter `this` pointer).

This solves it by sharing the promise and function with the implicit
converted function, rather than requiring that the `this` pointer remain
alive.