Revert Packet storage to a variant
Having a concrete span member broke the implicit copy and move
constructors because upon a copy of a Packet owning its own data the
vector would get copied but the span would still point at the old
object.

Although this could be fixed with explicit copy/move constructors to
untangle the references, at that point it is easier to simply go back to
using a variant that doesn't have these problems.