Remove dangerous `= nullptr` default from send keep-alives
Various stream and datagram methods take a bspan (or string_view) with a
keepalive, where the keepalive defaults to nullptr.  This is error-prone
as the keep-alive is *required* to keep the span alive, and if you don't
want or need it, you should need to be explicit about that by passing an
explicit nullptr.  Otherwise code like this is silently broken:

    std::vector<std::byte> stuff;
    // insert things into stuff
    stream->send(stuff);

and that is much too dangerous for a public API to allow.  With this
change that above no longer compiles and you need to change either to
move the vector into the send call (at which point we make the keepalive
for you) or provide some other object that allows libquic to guarantee
that the data will remain valid.