Make Endpoint destruction shut down connections
Without this we can end up not calling halt_events on the connection,
which can result in connection events still firing after the Connections
owned by the Endpoint have been destroyed.

This also fixes an issue triggered in the test suite's packet_delayer
where, because of the UDPSocket off the loop thread, a segfault could
happen because of UDPSocket being destroyed while being actively used in
the loop thread.